sequence cache: /prefix/txn_id/seq_num -> {}
it is better to put the sequence number into the key and deal with
duplicate entries than avoiding them at the cost of more unmarshalling,
especially during a merge. The sequence number is encoded in decreasing
order and GetSequence uses a one-element scan to check for the entry
(with mild optimizations).

@bdarnell pointed out that this might somewhat decrease RocksDB per-
formance since point lookups can use a bloom filter more effectively,
but there are custom facilities in RocksDB that should allow us to
tune the bloom filter for the sequence cache span should it show up
in performance tests.

+++ storage/replica.go
@@ -1031,7 +1031,7 @@ func (r *Replica) applyRaftCommandInBatch(ctx context.Context, index uint64, ori
 	if ba.IsWrite() && ba.Txn != nil {
 		if sequence, readErr := r.sequence.GetSequence(btch, ba.Txn.ID); readErr != nil {
 			return btch, nil, nil, newReplicaCorruptionError(util.Errorf("could not read from sequence cache"), readErr)
-		} else if sequence >= int64(ba.Txn.Sequence) {
+		} else if sequence >= ba.Txn.Sequence {
 			if log.V(1) {
 				log.Infoc(ctx, "found sequence cache entry for %s@%d", ba.Txn.Short(), ba.Txn.Sequence)
 			}
@@ -1094,7 +1094,7 @@ func (r *Replica) applyRaftCommandInBatch(ctx context.Context, index uint64, ori
 		}
 		// Only transactional requests have replay protection.
 		if ba.Txn != nil {
-			if putErr := r.sequence.PutSequence(btch, ba.Txn.ID, int64(ba.Txn.Sequence), err); putErr != nil {
+			if putErr := r.sequence.PutSequence(btch, ba.Txn.ID, ba.Txn.Sequence, err); putErr != nil {
 				// TODO(tschottdorf): ReplicaCorruptionError.
 				log.Fatalc(ctx, "putting a sequence cache entry in a batch should never fail: %s", putErr)
 			}
